Refactor TomlDate class to use java.util.Date
This commit refactors the TomlDate class to use the java.util.Date class instead of a String for storing the value of the TOML date. This change improves the functionality and readability of the class. The toTomlString() method is also updated to convert the Date object to a string representation.