Add README.md
This commit has added a README.md file to the project. This file contains instructions on how to run and use the project. It also provides information about the status of the project and its features. Detailed information and instructions on how to get started can be found within this file.